backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
boto3
pypdf
mangum
groq